fqdn_rand: don't stomp over the ruby kernel rng
fqdn_rand calls srand with the md5 of the fqdn. This means that from
from this point forwards rand() isn't as random as it might have been.

This attempt at a fix needs 1.9.2 in order to use the Random class added in
that release.  This will need backporting in some manner.

https://github.com/marcandre/backports/tree/master/lib/backports/1.9.2/random
may be the answer to supporting < 1.9.2